feat: Implement invest page and investment detail page
This commit introduces the new investment opportunities feature.

- Creates the `/invest` page to display a list of active investment opportunities fetched from Firestore.
- Implements a grid of investment cards with a loading skeleton state, an empty state, and error handling via toasts.
- Creates the `/invest/[slug]` dynamic detail page to show comprehensive information about a single investment.
- The detail page is built as a client-side component to ensure compatibility with Firestore's client-side data fetching and includes a loading state and 404 handling.
- Adds an "Invest Now" modal with a form for submitting an investment interest, including validation with Zod and React Hook Form.
- Defines a clear `Investment` type and adds a `pct` utility function for calculating funding percentages.
